requirements.txt
Click>=8.1.0,<9.0.0
colorama>=0.4.0,<1.0.0
pyautogen>=0.8.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
azure-ai-inference>=1.0.0b9
